MISRA: fix rule 10.8

+++ source/FreeRTOS_IP_Utils.c
@@ -677,7 +677,7 @@ uint16_t usGenerateProtocolChecksum( uint8_t * pucEthernetBuffer,
             usChecksum = ( uint16_t )
                          ( ~usGenerateChecksum( usChecksum,
                                                 ipPOINTER_CAST( const uint8_t *, &( pxIPPacket->xIPHeader.ulSourceIPAddress ) ),
-                                                ( size_t ) ( ( 2U * ipSIZE_OF_IPv4_ADDRESS ) + ulLength ) ) );
+                                                ( size_t ) ( ( 2U * ( size_t ) ipSIZE_OF_IPv4_ADDRESS ) + ulLength ) ) );
             /* Sum TCP header and data. */
         }
